Use battery name generated from glob pattern instaed of hard coding the name.
Previously, the battery name was hard-coded to BAT1 which mightn't be the battery designation for every device, so using the battery name generated from glob pattern while opening the file charge_full.

Additionally, it can be further improved to collect separate data for each batteries in case of multiple batteries (if required).